AV-1569: fix drupal component auth state issues

+++ ckan/scripts/init_filesystems.sh
@@ -6,6 +6,8 @@ if [[ "${DEV_MODE}" == "true" ]]; then
     echo "init_filesystems - skipping because DEV_MODE = 'true' ..."
 else
   # migrate data if migration fs is mounted to known path
+  # NOTE: this is a bad idea, because EFS->EFS rsync perf is so bad!
+  #       it's left here as an example, can be deleted later...
   if [[ -d "/mnt/ytp_files/ckan" ]]; then
     echo "init_filesystems - migrating data from '/mnt/ytp_files/ckan' to '${DATA_DIR}' ..."
     rsync -au --chown=ckan:ckan /mnt/ytp_files/ckan/ ${DATA_DIR}
@@ -15,9 +17,13 @@ else
     ls -lah ${DATA_DIR}
   fi
 
-  # init mounted filesystems (ECS Fargate EFS limitation forces this approach)
-  echo "init_filesystems - initializing '${DATA_DIR}' & '${WWW_DIR}/resources' ..."
-  rsync -au ${BASE_DIR}/data/ ${DATA_DIR} &
-  rsync -au --delete ${BASE_DIR}/resources/ ${WWW_DIR}/resources &
-  wait
+  # init mounted filesystems if not done or version updated (ECS Fargate EFS limitation forces this approach)
+  if [[ "$(cat ${DATA_DIR}/.fs-done)" != "$CKAN_IMAGE_TAG" ]]; then
+    echo "init_filesystems - initializing '${DATA_DIR}' & '${WWW_DIR}/resources' ..."
+    rsync -au ${BASE_DIR}/data/ ${DATA_DIR} &
+    rsync -au --delete ${BASE_DIR}/resources/ ${WWW_DIR}/resources &
+    wait
+    # set init flag to done
+    echo "$CKAN_IMAGE_TAG" > ${DATA_DIR}/.fs-done
+  fi
 fi
